+++ src/procurement_tools/cli/main.py
@@ -1,7 +1,7 @@
-from procurement_tools import get_entity
-from typing import Optional
-import typer
 import json
+from procurement_tools import USASpending, get_entity
+import typer
+from typing_extensions import Annotated
 
 app = typer.Typer()
 
@@ -13,9 +13,24 @@ def sam(uei: str):
     print(res.model_dump_json())
 
 
-@app.callback()
-def callback():
-    pass
+@app.command()
+def usaspending(
+    uei: str,
+    awards: Annotated[
+        bool,
+        typer.Option(
+            help="Get the latest awards for entity. If false, then gets the entity's profile data"
+        ),
+    ] = False,
+):
+    """Get JSON data about an entity from USASpending by providing a UEI"""
+
+    if awards:
+        res = USASpending.get_latest_recipient_awards(uei)
+        print(json.dumps(res, indent=2))
+    else:
+        res = USASpending.get_recipient_profile(uei)
+        print(json.dumps(res, indent=2))
 
 
 typer_click_object = typer.main.get_command(app)