+++ docs/CHANGELOG.rst
@@ -6,4 +6,6 @@ CHANGELOG
 
 New
 ~~~
-- A basic CLI. Right now, only :code:`fargo sam [UEI]` works.
+- A basic CLI.
+- A new USASpending method (:code:`USASpending.get_recipient_profile`), that gets a recipient profile from the USASpending.gov API
+- A new USASpending method (:code:`USASpending.get_latest_recipient_awards`), that gets the most recent 10 awards within the last 90 days from the USASpending.gov API